mostly harmless overflows, fixed

+++ src/modules/Image/encodings/x.c
@@ -605,7 +605,7 @@ static void image_x_encode_pseudocolor_1byte(INT32 args,
 	 x=img->xsize; 
 	 while (x--) 
 	 {
-	    b = ((*(s++))<<(32-bpp)) & 0xffffffff;
+	    b = (((unsigned long)*(s++))<<(32-bpp)) & 0xffffffff;
 	    bp = bpp;
 	    while (bp>8-bit)
 	    {
